+++ src/main/scala/uk/ac/cdrc/data/utility/text/Similarity.scala
@@ -30,14 +30,14 @@ trait NumberSpanExtractor {
   val spanPattern = "(\\d+)\\s*-\\s*(\\d+)".r
   val numPattern = "\\d+".r
 
-  def extractNumberSpan(s: String): Set[Int] = {
+  def extractNumberSpan(s: String): Set[Long] = {
     val numSpan = for {
       m <- (spanPattern findAllIn s).matchData
-      i <- m.group(1).toInt to m.group(2).toInt
+      i <- m.group(1).toLong to m.group(2).toLong
     } yield i
     val nums = for {
       m <- numPattern findAllIn (spanPattern replaceAllIn(s, " "))
-    } yield m.toInt
+    } yield m.toLong
     (numSpan ++ nums).toSet
   }
 
@@ -50,7 +50,7 @@ object NumberSpanDistance extends Similarity with NumberSpanExtractor{
     distance(numSetA, numSetB)
   }
 
-  def distance(a: Set[Int], b: Set[Int]): Float = {
+  def distance(a: Set[Long], b: Set[Long]): Float = {
     if (b.isEmpty)
       -1.0f // If the query do not contain any numbers then use -1f to mark it
     else
@@ -65,7 +65,7 @@ object NumbersOverlapDistance extends Similarity with NumberSpanExtractor {
     val numSetA = extractNumberSpan(a)
     distance(numSetA, numSetB)
   }
-  def distance(a: Set[Int], b: Set[Int]): Float = {
+  def distance(a: Set[Long], b: Set[Long]): Float = {
     val union: Float = (b | a).size
     if (union == 0)
       0
